new restructuring attempt
