--- src/backend/tracking/__init__.py
@@ -1,1 +1,0 @@
-# backend package loader file - placeholder
